requirements.txt
auger-python==0.1.35
Django==3.0.1
funcsigs==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
mock==2.0.0
pbr==5.1.1
pytz==2018.7
six==1.12.0
smmap2==2.0.5
psycopg2-binary==2.8.4
pycco
numpy
sklearn
